环信SDK更新

+++ hmjs/hmjs/Chat/Category/NSDate+Category.m
@@ -130,9 +130,9 @@
     if (timeInterval < 60) {
         return @"1分钟内";
 	} else if (timeInterval < 3600) {//1小时内
-        return [NSString stringWithFormat:@"%d分钟前", timeInterval / 60];
+        return [NSString stringWithFormat:@"%d分钟前", (int)(timeInterval / 60)];
 	} else if (timeInterval < 21600) {//6小时内
-        return [NSString stringWithFormat:@"%d小时前", timeInterval / 3600];
+        return [NSString stringWithFormat:@"%d小时前", (int)(timeInterval / 3600)];
 	} else if ([theDay isEqualToString:currentDay]) {//当天
 		[dateFormatter setDateFormat:@"HH:mm"];
         return [NSString stringWithFormat:@"今天 %@", [dateFormatter stringFromDate:self]];
